Push_swap : 125% with out leaks

+++ bonus/add_stack_bonus.c
@@ -40,8 +40,8 @@ static long	ft_atoi_bonus(const char *str)
 
 void	add_stack_a_bonus(t_stack_node **a, char **argv)
 {
-	int	i;
-	int	value;
+	int		i;
+	long	value;
 
 	i = 0;
 	while (argv[i])
